Use TxComposer in register delegate form
