+++ src/locales/en/common.json
@@ -419,7 +419,6 @@
   "The list shown only contains peers connected to the current Lisk Service node.": "The list shown only contains peers connected to the current Lisk Service node.",
   "The message can't contain whitespace at the beginning or end.": "The message can't contain whitespace at the beginning or end.",
   "The message signature has been canceled on your {{model}}": "The message signature has been canceled on your {{model}}",
-  "The minimum required balance to register is {{minBalance}} LSK": "The minimum required balance to register is {{minBalance}} LSK",
   "The provided amount is higher than available voting balance.": "The provided amount is higher than available voting balance.",
   "The signature is correct": "The signature is correct",
   "The signature is incorrect": "The signature is incorrect",
@@ -515,7 +514,6 @@
   "You cannot vote for this delegate": "You cannot vote for this delegate",
   "You don't have enough LSK in your account.": "You don't have enough LSK in your account.",
   "You don’t have any bookmarks yet.": "You don’t have any bookmarks yet.",
-  "You have already registered as a delegate.": "You have already registered as a delegate.",
   "You have cancelled the transaction on your hardware wallet.": "You have cancelled the transaction on your hardware wallet.",
   "You have now locked": "You have now locked",
   "You will be able to:": "You will be able to:",
